added cleanup step for output files
